bookings route

+++ app/controllers/bookings_controller.rb
@@ -1,7 +1,7 @@
 class BookingsController < ApplicationController
-    before_action :authorized
+  before_action :authorized
    
-    def index
+      def index
         render json: current_user_bookings
       end
     
@@ -9,9 +9,11 @@ class BookingsController < ApplicationController
         booking_id = params[:id]
         authorized(booking_id)
         booking = Booking.find(booking_id)
-        render json: booking
+        render json: { id: booking.id, user_id: booking.user_id, flight_id: booking.flight_id, user: booking.user, flight: booking.flight }
       end
-
+      
+      
+      
     
   def create
     flight = Flight.find(booking_params[:flight_id])
@@ -25,15 +27,26 @@ class BookingsController < ApplicationController
   end
 
   def destroy
-    booking = Booking.find_by(id: params[:id])
-    booking.destroy
-    head :no_content
+    flight_id = params[:flight_id]
+    user = current_user
+    booking = Booking.find_by(flight_id: flight_id, user_id: user.id)
+    if booking.present?
+      booking.destroy
+      render json: { message: "Booking successfully deleted" }, status: :ok
+    else
+      render json: { error: "Booking not found" }, status: :not_found
+    end
   end
-
+  
+  
+  
+  
     private
     def booking_params
         params.require(:booking).permit(:flight_id)
-      end
+    end
+      
+      
 
 end
